requirements/requeriments.txt
gitpython
pandas
requests
pymysql
sqlalchemy
BeautifulSoup4